# perf(common): faster logs by caching intl.datetimeformat