@micalevisk 2023-01-07
can we split those two optizations into 2 separated commits