@H4ad 2023-01-11
Based on https://github.com/jmcdo29/ogma/pull/1493, If we use the library [sonic-boom](https://github.com/pinojs/sonic-boom), I think we can achieve 200k~300k/ops.

In ogma, we went from 100k to 400k op/s.

Sonic boom is part of PinoJS and it buffers the logs, then sends to directly to file descriptor, using this approach is faster than writing directly to `process.stdout`.

Do you think it's worth optimizing?