@kamilmysliwiec 2023-02-01
Sounds good @H4ad. Would you like to create a separate PR for this?